Add files via upload
QuestieTracker is where the main changes are, added comments. Did not add a 'setting' for it, so that still needs doing (currently on 'if true then' override of any other setting).

Lines
1-4 - Include ZoneHelper library and seems required to register for an event? Dunno how it works
47-97 - Registers for the 'entering world' event and starts timer, also a couple subfunctions for distance calculation needed both by timer and main function
569-640 - Main function of addon to find the closest objective of each quest
684-704 - Calls the main function to perform the sort OnUpdate

ZoneHelper is a lib that was in BQT that I copied the way they used it. Had to add to embeds.xml of course.

I am very out of practice at WoW coding. This is very quick and dirty and rough, but it works. I'm sure the code could be very much cleaned to go in line with the rest of the code style, reduce duplication of functions, etc.