Add files via upload
So I found the function QuestieMap:GetNearestQuestSpawn which simplified things. Used that function instead of BQT's code.

That simplifies the code block for the GetDistanceToClosestObjective function on line 569.

Also, on line 964-968, I added a feature I wanted as well; whenever the tracker updates, put a tomtom arrow at the closest quest. In the main branch, this should be created as an option that's only available when sorting by proximity.